refector: logging 책임 분리
